Scanning code to separate module #110
